ofmcc-4619 - remove publicService
