+++ frontend/src/services/lookupService.js
@@ -10,4 +10,14 @@ export default {
       throw e
     }
   },
+
+  async getSystemMessages() {
+    try {
+      const response = await ApiService.apiAxios.get(ApiRoutes.SYSTEM_MESSAGES)
+      return response.data
+    } catch (error) {
+      console.log(`Failed to get system messages - ${error}`)
+      throw error
+    }
+  },
 }
